Remove template code from cpp file
